Revise assertTypedEqual to actually compare scalars at the leaves.
Also fixed the code to properly extract the fieldName and type.  The
code from before seems like it was written as if it were a `for val
of` loop rather than the actual `for key in` loop that it was using.

Finally, when an assertion failure occurs, this version also prints
out the original (outermost) inputs prior to the recursive traversal
of the tree-structure for the input type.